requirements.txt
urllib3==1.25.9
aiogoogle==1.1.1
aiohttp==3.7.3
arrow==0.17.0
beautifulsoup4==4.9.3
fastapi==0.63.0
google_api_python_client==1.12.8
google_auth_oauthlib==0.4.2
gspread==3.6.0
gspread_dataframe==3.2.0
jose==1.0.0
marshmallow==3.10.0
motor==2.3.0
numpy==1.19.5
pandas==1.2.0
passlib==1.7.4
pydantic==1.7.3
pymongo==3.11.2
python_dateutil==2.8.1
requests_toolbelt==0.9.1
secrets==1.0.2
starlette==0.14.1
tabulate==0.8.7
umongo==2.3.0
uvloop==0.14.0
